Refine: Add separate post_actions module
